requirements.txt
fastapi
uvicorn
mysqlclient
passlib[bcrypt]
python-jose[cryptography]
slowapi
sqlalchemy
mysql-connector-python
pandas
openai==0.28
ultralytics
